move example to slash_options.py
